[grid.rect|tria] minor tweaks

+++ src/pymor/grid/rect.py
@@ -93,7 +93,7 @@ class Rect(ISimpleAffineGrid):
         assert codim <= subentity_codim <= self.dim, CodimError('Invalid subentity codimensoin')
         if codim == 0:
             if subentity_codim == 0:
-                return np.arange(self.size(0))[:, np.newaxis]
+                return np.arange(self.size(0), dtype='int32')[:, np.newaxis]
             else:
                 return self.__subentities[subentity_codim - 1]
         else:
